backend/agent-service/app/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
pypdf
faiss-cpu
langchain
langchain-core
langchain-community
langchain-huggingface
langgraph
playwright
huggingface_hub
requests
beautifulsoup4

numpy
tqdm

